[FLINK-25141][connector/elasticsearch] Add sink parallelism option
